requirements.txt
streamlit>=1.33.0
pandas>=2.1.0
numpy>=1.25.0
Pillow>=10.0.0
opencv-python-headless==4.8.0.74
fer==22.4.0
